// src/types/corporation.ts
export interface Corporation {
  id: number
  sku_corporation: string
  corporation_name: string
  corporation_name_mini: string
  ceo: string
  phone: string
  email: string
  country: string
  state: string
  town: string
  adresse: string
  employee_capacity: string
  send_message_task: string
  logo_url: string
  login_cover_image?: string
  login_title?: string
  login_subtitle?: string
  created_at: string
  updated_at: string
}